src/backend/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
pydantic
jinja2
python-multipart
httpx